requirements.txt
Django==6.0.2
djangorestframework==3.15.2
django-cors-headers==4.6.0
dj-database-url==2.3.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
celery==5.4.0
redis==5.2.1
django-redis==5.4.0
gunicorn==23.0.0
twilio==9.4.3
Pillow==11.1.0
matplotlib==3.10.0
seaborn==0.13.2
